requirements.txt
playwright==1.51.0
pytest==8.3.5
pytest-playwright==0.7.0
httpx==0.28.1
python-dotenv==1.1.0
